requirements.txt
# Core
openai
python-dotenv
pydantic

# FastAPI Backend
fastapi
uvicorn[standard]
python-multipart

# Streamlit Frontend
streamlit
requests

# Vector Store
qdrant-client

# LlamaIndex
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-cloud-services

# Embeddings
sentence-transformers

# Evaluation
ragas

# Utilities
Pillow